@@ -9,14 +9,14 @@ apply plugin: 'kotlin-kapt'
 
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.troy.playgroundkotlin"
         minSdkVersion 23
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     dataBinding {
         enabled = true
@@ -30,8 +30,8 @@ android {
 }
 
 ext {
-    android_support_version = '27.1.1'
-    dagger_version = '2.15'
+    android_support_version = '28.0.0'
+    dagger_version = '2.21'
     viewmodel_version = '1.1.1'
     rxjava_version = '2.1.6'
     rxandroid_version = '2.0.1'
@@ -41,13 +41,13 @@ ext {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "com.android.support:appcompat-v7:${android_support_version}"
-    implementation "com.android.support:recyclerview-v7:${android_support_version}"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
     //     Dagger 2 - dependency injection
     def daggerCompiler = "com.google.dagger:dagger-compiler:$dagger_version"
@@ -63,8 +63,8 @@ dependencies {
     annotationProcessor "com.google.dagger:dagger-android-processor:$dagger_version"
 
     // View Model
-    implementation "android.arch.lifecycle:extensions:$viewmodel_version"
-    annotationProcessor "android.arch.lifecycle:compiler:$viewmodel_version"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    annotationProcessor 'androidx.lifecycle:lifecycle-compiler:2.2.0'
 
     // RxJava
     implementation "io.reactivex.rxjava2:rxjava:${rxjava_version}"
