@@ -1,7 +1,7 @@
 package com.troy.playgroundkotlin
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import com.troy.playgroundkotlin.core.base.BaseFragment
 import dagger.android.support.DaggerAppCompatActivity
 
