@@ -1,6 +1,6 @@
 package com.troy.playgroundkotlin.core.utility
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 
